Use domain wide delegation for getSpreadsheetData
