+++ src/components/FirecallItems/EinsatzAction.ts
@@ -1,9 +1,9 @@
 'use server';
 
 import { google } from 'googleapis';
-import { GoogleAuth, JWT } from 'googleapis-common';
 import moment from 'moment';
 import { parseTimestamp } from '../../common/time-format';
+import { createWorkspaceAuth } from '../../server/auth/workspace';
 import { Firecall } from '../firebase/firestore';
 
 /**
@@ -17,28 +17,9 @@ export async function copyFirecallSheet(firecall: Firecall): Promise<string> {
     `copying firecall sheet template for ${firecall.id} ${firecall.name}`
   );
 
-  let auth: GoogleAuth | JWT = new google.auth.GoogleAuth({
-    scopes: SCOPES,
-  });
-
-  if (
-    process.env.GOOGLE_SERVICE_ACCOUNT &&
-    process.env.EINSATZMAPPE_IMPERSONATION_ACCOUNT
-  ) {
-    const serviceAccount = JSON.parse(process.env.GOOGLE_SERVICE_ACCOUNT);
-    auth = new google.auth.JWT({
-      email: serviceAccount.client_email,
-      key: serviceAccount.private_key,
-      keyId: serviceAccount.private_key_id,
-      projectId: serviceAccount.project_id,
-      clientId: serviceAccount.client_id,
-      scopes: SCOPES,
-      subject: process.env.EINSATZMAPPE_IMPERSONATION_ACCOUNT,
-    });
-  }
-
   const tmstp = moment(parseTimestamp(firecall.alarmierung));
 
+  const auth = createWorkspaceAuth(SCOPES);
   const drive = google.drive({ version: 'v3', auth });
 
   const parentFolder = process.env.EINSATZMAPPE_SHEET_FOLDER || '';
